Users/Groups now working perfectly (group management in admin); Container/Images should work too expect clone/start

+++ wui/ipynbsrv/wui/signals/shares.py
@@ -2,43 +2,45 @@ import os.path
 import shutil
 import stat
 from django.conf import settings
-from django.db.models.signals import post_delete, post_save, pre_delete, pre_save
+from django.db.models.signals import post_delete, post_save
 from django.dispatch import receiver
 from ipynbsrv.wui.models import LdapGroup, Share
 from ipynbsrv.wui.signals.signals import *
 from ipynbsrv.wui.tools import Filesystem
 
 
-@receiver(post_share_created)
+@receiver(share_created)
 def create_share_directory(sender, share, **kwargs):
     """
     Signal receiver that creates the filesystem directory for a created share.
     """
     if settings.DEBUG:
         print "create_share_directory receiver fired"
-    # create the directory
-    path = os.path.join(settings.SHARE_ROOT, share.name)
-    Filesystem.ensure_directory(path)
-    # set owner and permissions
-    ldap_group = LdapGroup.objects.get(name=settings.SHARE_GROUP_PREFIX + share.name)
-    os.chown(path, -1, ldap_group.id)
-    os.chmod(path, stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR | stat.S_IRGRP | stat.S_IWGRP | stat.S_IXGRP | stat.S_ISGID)
+    if share is not None:
+        # create the directory
+        path = os.path.join(settings.SHARE_ROOT, share.name)
+        Filesystem.ensure_directory(path)
+        # set owner and permissions
+        ldap_group = LdapGroup.objects.get(name=settings.SHARE_GROUP_PREFIX + share.name)
+        os.chown(path, -1, ldap_group.id)
+        os.chmod(path, stat.S_IRUSR | stat.S_IWUSR | stat.S_IXUSR | stat.S_IRGRP | stat.S_IWGRP | stat.S_IXGRP | stat.S_ISGID)
 
 
-@receiver(post_share_deleted)
+@receiver(share_deleted)
 def delete_share_directory(sender, share, **kwargs):
     """
     Signal receiver that deletes the filesystem directory for a removed share.
     """
     if settings.DEBUG:
         print "delete_share_directory receiver fired"
-    # remove the directory
-    shutil.rmtree(path=os.path.join(settings.SHARE_ROOT, share.name), ignore_errors=not settings.DEBUG)
-    # make sure the group is removed too
-    share.group.delete()
+    if share is not None:
+        # remove the directory
+        shutil.rmtree(path=os.path.join(settings.SHARE_ROOT, share.name), ignore_errors=not settings.DEBUG)
+        # make sure the group is removed too
+        share.group.delete()
 
 
-@receiver(post_share_modified)
+@receiver(share_modified)
 def share_modified_handler(sender, share, fields, **kwargs):
     if settings.DEBUG:
         print "share_modified_handler receiver fired"
@@ -50,31 +52,12 @@ def share_modified_handler(sender, share, fields, **kwargs):
 
 @receiver(post_delete, sender=Share)
 def post_delete_handler(sender, instance, **kwargs):
-    post_share_deleted.send(sender=sender, share=instance, kwargs=kwargs)
-    share_deleted.send(sender=sender, share=instance, action='post_delete', kwargs=kwargs)
-
-
-@receiver(pre_delete, sender=Share)
-def pre_delete_handler(sender, instance, **kwargs):
-    pre_share_deleted.send(sender=sender, share=instance, kwargs=kwargs)
-    share_deleted.send(sender=sender, share=instance, action='pre_delete', kwargs=kwargs)
+    share_deleted.send(sender=sender, share=instance, kwargs=kwargs)
 
 
 @receiver(post_save, sender=Share)
 def post_save_handler(sender, instance, **kwargs):
-    if kwargs['created']:
-        post_share_created.send(sender, share=instance, kwargs=kwargs)
-        share_created.send(sender, share=instance, action='post_save', kwargs=kwargs)
+    if 'created' in kwargs:
+        share_created.send(sender, share=instance, kwargs=kwargs)
     else:
-        post_share_modified.send(sender, share=instance, fields=kwargs['update_fields'], kwargs=kwargs)
-        share_modified.send(sender, share=instance, fields=kwargs['update_fields'], action='post_save', kwargs=kwargs)
-
-
-@receiver(pre_save, sender=Share)
-def pre_save_handler(sender, instance, **kwargs):
-    if kwargs['created']:
-        pre_share_created.send(sender, share=instance, kwargs=kwargs)
-        share_created.send(sender, share=instance, action='pre_save', kwargs=kwargs)
-    else:
-        pre_share_modified.send(sender, share=instance, fields=kwargs['update_fields'], kwargs=kwargs)
-        share_modified.send(sender, share=instance, fields=kwargs['update_fields'], action='pre_save', kwargs=kwargs)
+        share_modified.send(sender, share=instance, fields=kwargs['update_fields'], kwargs=kwargs)
